requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4~=4.13.5
behave~=1.3.2
behave-html-formatter==0.9.10
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
cucumber-expressions==18.0.1
cucumber-tag-expressions==6.2.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.8.2
geventhttpclient==2.3.4
gherkin-official==32.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
locust==2.37.14
locust-cloud==1.26.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.4
msgpack==1.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.2
openai==2.3.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parse==1.20.2
parse_type==0.6.4
parso==0.8.5
pickleshare==0.7.5
pillow~=11.3.0
pip-check-reqs==2.5.5
pipreqs==0.4.13
platformdirs==4.4.0
pluggy==1.5.0
plyer==2.1.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.3.4
pytest-check==2.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pywin32==311; sys_platform == 'win32'
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests~=2.32.5
responses==0.25.8
rpds-py==0.27.1
rsa==4.9.1
selenium~=4.28.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
softest==1.2.0.0
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
tenacity==9.1.2
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarg==0.1.9
zope.event==5.1.1
zope.interface==7.2
google-generativeai==0.8.5